remove duplication in the localization tests
